requirements.txt
baidu-aip==2.2.18.0
certifi==2019.9.11
chardet==3.0.4
Django==1.10.2
djangorestframework==3.10.3
idna==2.8
Markdown==2.5.2
Pygments==2.0.1
PyMySQL==0.8.0
pytz==2019.3
qiniu==7.2.6
requests==2.22.0
urllib3==1.25.6
